requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
cloudinary>=1.35.0
